feat: create @mlx-node/cli package with unified `mlx` CLI
Migrate standalone scripts (download-model, convert-model, download-gsm8k)
into a proper @mlx-node/cli subpackage with subcommand routing:
  mlx download model, mlx download dataset, mlx convert
