+++ packages/cli/src/commands/download-dataset.ts
@@ -1,13 +1,11 @@
-import { mkdir, readdir, stat, copyFile } from 'node:fs/promises';
-import { existsSync } from 'node:fs';
+import { readdir, stat, copyFile } from 'node:fs/promises';
+import { parseArgs } from 'node:util';
 import { join, dirname, resolve } from 'node:path';
-import { fileURLToPath } from 'node:url';
 import { snapshotDownload } from '@huggingface/hub';
 
 import { convertParquetToJsonl } from '@mlx-node/core';
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = dirname(__filename);
+import { ensureDir } from '../utils.js';
 
 const DEFAULT_DATASET = 'openai/gsm8k';
 const DEFAULT_REVISION = 'main';
@@ -16,6 +14,25 @@ const FILE_SPECS = [
   { output: 'train.jsonl', parquetPrefix: 'train-' },
   { output: 'test.jsonl', parquetPrefix: 'test-' },
 ];
+
+function printHelp(): void {
+  console.log(`
+Download a dataset from HuggingFace
+
+Usage:
+  mlx download dataset [options]
+
+Options:
+  -d, --dataset <name>    HuggingFace dataset name (default: ${DEFAULT_DATASET})
+  -r, --revision <rev>    Dataset revision (default: ${DEFAULT_REVISION})
+  -o, --output <dir>      Output directory (default: data/gsm8k)
+  -h, --help              Show this help message
+
+Examples:
+  mlx download dataset
+  mlx download dataset --dataset openai/gsm8k --output data/gsm8k
+`);
+}
 
 async function findFirstMatch(
   root: string,
@@ -37,23 +54,49 @@ async function findFirstMatch(
   return null;
 }
 
-async function ensureDir(path: string): Promise<void> {
-  if (!existsSync(path)) {
-    await mkdir(path, { recursive: true });
-  }
-}
+export async function run(argv: string[]) {
+  const { values: args } = parseArgs({
+    args: argv,
+    options: {
+      dataset: {
+        type: 'string',
+        short: 'd',
+        default: process.env.GSM8K_DATASET ?? DEFAULT_DATASET,
+      },
+      revision: {
+        type: 'string',
+        short: 'r',
+        default: process.env.GSM8K_REVISION ?? DEFAULT_REVISION,
+      },
+      output: {
+        type: 'string',
+        short: 'o',
+        default: process.env.GSM8K_OUTPUT_DIR ?? DEFAULT_OUTPUT_DIR,
+      },
+      help: {
+        type: 'boolean',
+        short: 'h',
+        default: false,
+      },
+    },
+  });
 
-async function main() {
-  const dataset = process.env.GSM8K_DATASET ?? DEFAULT_DATASET;
-  const revision = process.env.GSM8K_REVISION ?? DEFAULT_REVISION;
-  const outputDir = process.env.GSM8K_OUTPUT_DIR ? resolve(process.env.GSM8K_OUTPUT_DIR) : DEFAULT_OUTPUT_DIR;
+  if (args.help) {
+    printHelp();
+    return;
+  }
+
+  const dataset = args.dataset!;
+  const revision = args.revision!;
+  const outputDir = resolve(args.output!);
 
   console.log(`Downloading ${dataset}@${revision} snapshot from Hugging Face…`);
 
+  const cacheDir = join(process.cwd(), '.cache', 'huggingface');
   const snapshotPath = await snapshotDownload({
     repo: { type: 'dataset', name: dataset },
     revision,
-    cacheDir: join(__dirname, '..', '.cache', 'huggingface'),
+    cacheDir,
   });
 
   console.log(`Snapshot available at ${snapshotPath}`);
@@ -93,8 +136,3 @@ async function main() {
   console.log('Done.');
   console.log(`Dataset files stored under ${outputDir}`);
 }
-
-main().catch((error) => {
-  console.error('[download-gsm8k] failed:', error);
-  process.exitCode = 1;
-});
